third_party/blink/web_tests/external/wpt/css/cssom/getComputedStyle-pseudo-picker-expected.txt

This is a testharness.js-based test.
[FAIL] ::picker(select) on non-select element resolves the style
  assert_equals: expected "rgb(0, 0, 255)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
